+++ src/main.cpp
@@ -1,8 +1,11 @@
-#include <Geode/Geode.hpp>
+#include "popups/CustomPopupLayer.hpp"
 
-using namespace geode::prelude;
+#include <Geode/Geode.hpp>
+#include <Geode/Loader.hpp>
 
 #include <Geode/modify/EditorPauseLayer.hpp>
+
+using namespace geode::prelude;
 
 class $modify(EditorPause, EditorPauseLayer)
 {
@@ -10,7 +13,24 @@ class $modify(EditorPause, EditorPauseLayer)
 	{
 		if (EditorPauseLayer::init(p0))
 		{
-			// code here
+			auto actionsMenu = this->getChildByID("actions-menu");
+			auto smallActionsMenu = this->getChildByID("small-actions-menu");
+			auto togglesMenu = this->getChildByID("options-menu");
+
+			actionsMenu->removeMeAndCleanup();
+			smallActionsMenu->removeMeAndCleanup();
+
+			togglesMenu->removeAllChildrenWithCleanup(true);
+
+			auto newTogglesBtnSprite = CCSprite::createWithSpriteFrameName("GJ_plainBtn_001.png");
+			newTogglesBtnSprite->setScale(0.875);
+
+			auto newTogglesBtn = CCMenuItemSpriteExtra::create(
+				newTogglesBtnSprite,
+				this,
+				menu_selector(EditorPause::onNewToggles));
+
+			togglesMenu->addChild(newTogglesBtn);
 
 			return true;
 		}
@@ -18,5 +38,10 @@ class $modify(EditorPause, EditorPauseLayer)
 		{
 			return false;
 		};
+	};
+
+	void onNewToggles(CCObject * sender)
+	{
+		CustomPopupLayer::create()->show();
 	};
 };