+++ changelog.md
@@ -1,2 +1,6 @@
 # 1.0.0
-- Edit this file to change your mod's changelog.
+#### Initial release.
+
+- ![Add Icon](frame:GJ_completesIcon_001.png&scale:0.5) | **Added** button in level editor pause menu for editor options
+- ![Remove Icon](frame:GJ_deleteIcon_001.png&scale:0.5) | **Removed** actions menu from level editor pause menu
+- ![Remove Icon](frame:GJ_deleteIcon_001.png&scale:0.5) | **Removed** small actions menu from level editor pause menu